Move chargen changes from to recreation
The RAM changes were inadvertantly made in the schematic capture
directory, which should serve as a reference.  This is implemented as a
separate directory rather than a branch because this documents and
clarifies the original schematic, rather than simply representing a
different revision.

The RAM changes on page 6 (chargen) were added to the recreation, and
the changes were reverted from the schematic capture.